Intro Screen, Navigation Screen, Basic image filters
Added an intro screen with our logo, a navigation screen for going between available areas (Currently only has Goodricke) and have added a system that allows changing the colour and alpha of ImageActors. This filtering is not yet in AnimatedActors as it didn't have any foreseeable use. Also fixed a bug in ResourceManager which made it only load the last image.